requirements.txt
bcrypt
flask_sqlalchemy
flask_session
python-dotenv
flask
redis
psycopg2